Move MongoDB classes to different packages
